initial ML model selection prototype
This isn't final by any means, as a lot of the models are classification and we need to use regression for the travel time output, and I'm not sure which scoring mechanism is best for regression, but it's a start.
Also converted the training data structure to a list of dictionaries that are ordered the same as the training data.