Make the ML trained model usable for the simulator
New file, predictML.py implements a function predict() that takes an array of the 16 variables used to train the ML model to make a prediction based on the 16 variables given. Includes an example. The 'training_samples_10k.csv' needs to be in the same folder as the predictML.py file as it does the training when you import predictML.py.